+++ lib_acl/src/xml/acl_xml_util.c
@@ -303,7 +303,7 @@ ACL_XML_NODE *acl_xml_create_node(ACL_XML *xml, const char* tag,
 	if (text && *text) {
 		size_t len = strlen(text);
 
-		ACL_VSTRING_SPACE(node->text, len);
+		ACL_VSTRING_SPACE(node->text, (int) len);
 		acl_vstring_strcpy(node->text, text);
 	}
 	return node;
