+++ lib_acl_cpp/samples/aio/pkg_server/main.cpp
@@ -92,7 +92,7 @@ public:
 		DAT_HDR res_hdr;
 
 		// 将主机字节序转为网络字节序
-		res_hdr.len = (int) htonl(dat_len);
+		res_hdr.len = (int) htonl((unsigned long) dat_len);
 		ACL_SAFE_STRNCPY(res_hdr.cmd, "ok", sizeof(res_hdr.cmd));
 
 		// 异步写响应数据包: 数据头及数据体
