+++ lib_acl_cpp/samples/connect_manager/connect_pool.h
@@ -14,7 +14,7 @@ protected:
 
 private:
 	acl::string addr_;
-	int   count_;
+	size_t count_;
 	size_t idx_;
 	int   conn_timeout_;
 	int   rw_timeout_;
